pyproject.toml
[project]
name = "rmq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.1.0",
    "pika>=1.3.2",
    "python-dotenv>=1.2.1",
]
